mySQL fixes (support show create statements, QTBUG-4008)

+++ src/editor/tosyntaxanalyzermysql.cpp
@@ -63,7 +63,7 @@ toSyntaxAnalyzer::statementList toSyntaxAnalyzerMysql::getStatements(const QStri
     try
     {
         std::unique_ptr <SQLLexer::Lexer> lexer = LexerFactTwoParmSing::Instance().create("MySQLGuiLexer", "", "toSyntaxAnalyzerMysql");
-        lexer->setStatement(str.c_str(), str.length());
+        lexer->setStatement(str.c_str(), (int)str.length());
 
         SQLLexer::Lexer::token_const_iterator start = lexer->begin();
         start = lexer->findStartToken(start);
@@ -101,7 +101,7 @@ toSyntaxAnalyzer::statement toSyntaxAnalyzerMysql::getStatementAt(unsigned line,
     try
     {
         std::unique_ptr <SQLLexer::Lexer> lexer = LexerFactTwoParmSing::Instance().create("MysqlGuiLexer", "", "toCustomLexer");
-        lexer->setStatement(str.c_str(), str.length());
+        lexer->setStatement(str.c_str(), (int)str.length());
 
         SQLLexer::Lexer::token_const_iterator start = lexer->begin();
         start = lexer->findStartToken(start);
