+++ src/editor/tosyntaxanalyzeroracle.cpp
@@ -63,7 +63,7 @@ toSyntaxAnalyzer::statementList toSyntaxAnalyzerOracle::getStatements(const QStr
     try
     {
         std::unique_ptr <SQLLexer::Lexer> lexer = LexerFactTwoParmSing::Instance().create("OracleGuiLexer", "", "toCustomLexer");
-        lexer->setStatement(str.c_str(), str.length());
+        lexer->setStatement(str.c_str(), (int)str.length());
 
         SQLLexer::Lexer::token_const_iterator start = lexer->begin();
         start = lexer->findStartToken(start);
@@ -101,7 +101,7 @@ toSyntaxAnalyzer::statement toSyntaxAnalyzerOracle::getStatementAt(unsigned line
     try
     {
         std::unique_ptr <SQLLexer::Lexer> lexer = LexerFactTwoParmSing::Instance().create("OracleGuiLexer", "", "toCustomLexer");
-        lexer->setStatement(str.c_str(), str.length());
+        lexer->setStatement(str.c_str(), (int)str.length());
 
         SQLLexer::Lexer::token_const_iterator start = lexer->begin();
         start = lexer->findStartToken(start);
